:bug: Fix CSV reader error when header not first line
