+++ pystreamapi/loaders/__init__.py
@@ -1,5 +1,7 @@
-from pystreamapi.loaders.__csv_loader import csv
+from pystreamapi.loaders.__csv.__csv_loader import csv
+from pystreamapi.loaders.__json.__json_loader import json
 
 __all__ = [
-    'csv'
+    'csv',
+    'json'
 ]
