--- tests/assets/data.csv
@@ -1,3 +1,0 @@
-attr1,attr2
-1,2.0
-a,b
