--- tests/assets/data2.csv
@@ -1,2 +1,0 @@
-attr1;attr2
-1;2
